AS7-2102 Move fake JNDI hack to testsuite/shared (requested by ochaloup).
