+++ testsuite/integration/basic/src/test/resources/demos/fakejndi.sar/META-INF/jboss-service.xml
@@ -23,5 +23,5 @@
       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
       xsi:schemaLocation="urn:jboss:service:7.0 jboss-service_7_0.xsd">
 
-    <mbean name="jboss:name=test,type=fakejndi" code="org.jboss.as.test.smoke.embedded.demos.fakejndi.FakeJndi"/>
+    <mbean name="jboss:name=test,type=fakejndi" code="org.jboss.as.test.shared.util.fakejndi.FakeJndi"/>
 </server>
