+++ testsuite/integration/smoke/src/test/java/org/jboss/as/test/smoke/embedded/mgmt/datasource/DataSourceOperationsUnitTestCase.java
@@ -58,7 +58,13 @@ import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.container.test.api.RunAsClient;
 import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.as.arquillian.container.TunneledMBeanServerConnection;
-import org.jboss.as.test.smoke.embedded.demos.fakejndi.FakeJndi;
+import org.jboss.as.connector.subsystems.datasources.DataSourcesExtension.NewDataSourceSubsystemParser;
+import org.jboss.as.connector.subsystems.datasources.ModifiableXaDataSource;
+import org.jboss.as.connector.subsystems.datasources.Namespace;
+import org.jboss.as.controller.client.ModelControllerClient;
+import org.jboss.as.controller.persistence.SubsystemMarshallingContext;
+import org.jboss.as.protocol.StreamUtils;
+import org.jboss.as.test.shared.util.fakejndi.FakeJndi;
 import org.jboss.as.test.smoke.modular.utils.PollingUtils;
 import org.jboss.as.test.smoke.modular.utils.ShrinkWrapUtils;
 import org.jboss.dmr.ModelNode;
