+++ testsuite/integration/smoke/src/test/resources/demos/fakejndi.sar/META-INF/jboss-service.xml
@@ -20,8 +20,8 @@
   ~ 02110-1301 USA, or see the FSF site: http://www.fsf.org.
   -->
 <server xmlns="urn:jboss:service:7.0"
-      xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-      xsi:schemaLocation="urn:jboss:service:7.0 jboss-service_7_0.xsd">
+        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:schemaLocation="urn:jboss:service:7.0 jboss-service_7_0.xsd">
 
-    <mbean name="jboss:name=test,type=fakejndi" code="org.jboss.as.test.smoke.embedded.demos.fakejndi.FakeJndi"/>
+    <mbean name="jboss:name=test,type=fakejndi" code="org.jboss.as.test.shared.util.fakejndi.FakeJndi"/>
 </server>
